put new functionality in a package extention
