+++ Project.toml
@@ -19,15 +19,17 @@ WoodburyMatrices = "efce3f68-66dc-5838-9240-27a6d6f5f9b6"
 
 [weakdeps]
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
+ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 
 [extensions]
 InterpolationsUnitfulExt = "Unitful"
+InterpolationsForwardDiffExt = "ForwardDiff"
 
 [compat]
 Adapt = "2, 3, 4.0"
 AxisAlgorithms = "0.3, 1"
 ChainRulesCore = "0.10, 1.0, 1.2, 1.3"
-ForwardDiff = "1.0.1"
+ForwardDiff = "0.10, 1.0"
 OffsetArrays = "0.10, 0.11, 1.0.1"
 Ratios = "0.3, 0.4"
 Requires = "1.1"
